PLF-8084 : change log about missing FCM configuration to warn level
The log message telling that the Firebase Cloud Messaging configuration file is missing was in error level.
Since activating push notifications is not mandatory (customers can decide to not use this feature), the log level has been changed to warn level and updated to make it clear that Push Notifications feature will not work in this case.